dedupe participants
